Fix issue#3
Issue was "paths with spaces aren't handled properly".
In bm_bash, had to surround the output of bm with double-quotes.
What if the original filename has double-quotes in it? So, had
to escape the double-quotes in shellquote function in bm_util.
Also, during debugging had to repeatedly source ~/.bashrc after
running ./install.sh. So added that instruction to the user also.